Made reviews admin panel and changed delete methods in other admin panels

+++ src/main/java/com/epam/furniturestoreapp/controller/AdminOrdersController.java
@@ -38,9 +38,11 @@ public class AdminOrdersController {
         return "orders-admin";
     }
 
-    @DeleteMapping("/delete/{id}")
+    @DeleteMapping("/{id}")
     public String deleteOrderAdmin(@PathVariable Long id){
-        orderTableService.deleteById(id);
+        if(orderTableService.existById(id)){
+            orderTableService.deleteById(id);
+        }
         return "redirect:/orders-admin";
     }
 }
